[#704] Moved geometry matrix and inversions into dedicated classes
